+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v11.0.0 Breaking changes
+
+- Removed `Tooltip` component, use [@goodhood/components Tooltip](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `Slider` component, use [@goodhood/components Slider](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `Markdown` component, use [@goodhood/components Markdown](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `ContentHeader` component, use [@goodhood/components ContentHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `LinkHeader` component, use [@goodhood/components LinkHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `Dots` component, use [@goodhood/components Dots](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Removed `Accordion` component
+
 ## v10.0.0 Breaking changes
 
 - Removed `process` references